+++ home/dataResources/php/ajax.php
@@ -821,34 +821,6 @@ foreach ($requests as $request) {
             //$response->responseData->config->expandFilters = json_decode('{ "field":"id", "operator":"IN", "value":["1","1.1"] }');
             break;
 
-        case 'treeOld.load':
-            $tree = array();
-            $nodeId = $request->requestData->nodeId;
-
-            for ($i = 0; $i < 3; $i++) {
-                $node = new stdClass();
-
-                if ($nodeId === null) {
-                    $node->caption = 'Root ' . $i;
-                    $node->nodeId = $i;
-                    $node->leaf = $i <> 1;
-                } else {
-                    $node->caption = 'Knoten ' . $nodeId . '-' . $i;
-                    $node->nodeId = $nodeId . '-' . $i;
-                    $node->leaf = $i <> 1;
-                }
-                $tree[] = $node;
-
-            }
-
-            // Verzögerung um Lademaske anzuzeigen
-            if ($nodeId !== null) {
-                sleep(rand(0, 2));
-            }
-
-            $response->responseData->tree = $tree;
-            break;
-            
         default:
             $response->errorMsg = 'RemoteFn "' . $request->remoteFn . '" existiert nicht.';
     }